add local deployment of proxy

+++ sunholo/cli/run_proxy.py
@@ -1,5 +1,6 @@
 import subprocess
 import os
+import sys
 import signal
 import json
 
@@ -96,7 +97,7 @@ def save_proxies(proxies):
 
 
 
-def start_proxy(service_name, region, project, port=None):
+def start_proxy(service_name, region, project, port=None, local=False, app_type=None, app_folder=None, log_file=None):
     """
     Starts the gcloud proxy to the Cloud Run service and stores the PID.
 
@@ -114,34 +115,48 @@ def start_proxy(service_name, region, project, port=None):
 
     if not port:
         port = get_next_available_port(proxies, DEFAULT_PORT)
+    
 
-    command = [
-        "gcloud", "run", "services", "proxy", service_name,
-        "--region", region,
-        "--project", project,
-        "--port", str(port)
-    ]
-    with open(os.devnull, 'w') as devnull:
-        process = subprocess.Popen(command, stdout=devnull, stderr=devnull, preexec_fn=os.setpgrp)
-    
-    proxies[service_name] = {
-        "pid": process.pid,
-        "port": port
-    }
-    save_proxies(proxies)
-    
-    console.print(f"Proxy for [bold orange]'{service_name}'[/bold orange] setup complete on port {port}")
-    list_proxies()
+
+    if local:
+        start_local(service_name, port, app_type, app_folder, log_file)
+    else:
+        command = [
+            "gcloud", "run", "services", "proxy", service_name,
+            "--region", region,
+            "--project", project,
+            "--port", str(port)
+        ]
+        if log_file:
+            log_file_path = os.path.join(app_folder, f"{service_name}_log.txt")
+            with open(log_file_path, 'a') as logf:
+                process = subprocess.Popen(command, stdout=logf, stderr=logf, preexec_fn=os.setpgrp)
+        else:
+            log_file_path = "No log file specified"
+            with open(os.devnull, 'w') as devnull:
+                process = subprocess.Popen(command, stdout=devnull, stderr=devnull, preexec_fn=os.setpgrp)
+            
+        proxies[service_name] = {
+            "pid": process.pid,
+            "port": port,
+            "local": "No",
+            "logs": log_file_path
+        }
+        save_proxies(proxies)
+        
+        console.print(f"Proxy for [bold orange]'{service_name}'[/bold orange] setup complete on port {port}")
+        list_proxies()
 
     return f"http://127.0.0.1:{port}"
 
 
-def stop_proxy(service_name):
+def stop_proxy(service_name, stop_local=True):
     """
     Stops the gcloud proxy to the Cloud Run service using the stored PID.
 
     Args:
         service_name (str): Name of the Cloud Run service.
+        stop_local (bool): Whether to stop locally running services or not. Defaults to True.
     """
     proxies = clean_proxy_list()
 
@@ -149,16 +164,21 @@ def stop_proxy(service_name):
         print(f"No proxy found for service: {service_name}")
         return
 
-    pid = proxies[service_name]["pid"]
-    try:
-        os.kill(pid, signal.SIGTERM)
-        del proxies[service_name]
-        save_proxies(proxies)
-        console.print(f"Proxy for [bold orange]'{service_name}'[bold orange] stopped.")
-    except ProcessLookupError:
-        console.print(f"No process found with PID: {pid}")
-    except Exception as e:
-        console.print(f"[bold red]Error stopping proxy for {service_name}: {e}[/bold red]")
+    if not stop_local:
+        local = proxies[service_name]["local"]
+        if local != "No":
+            console.print(f"Not stopping local VAC running on: {local}")
+    else:
+        pid = proxies[service_name]["pid"]
+        try:
+            os.kill(pid, signal.SIGTERM)
+            del proxies[service_name]
+            save_proxies(proxies)
+            console.print(f"Proxy for [bold orange]'{service_name}'[bold orange] stopped.")
+        except ProcessLookupError:
+            console.print(f"No process found with PID: {pid}")
+        except Exception as e:
+            console.print(f"[bold red]Error stopping proxy for {service_name}: {e}[/bold red]")
     
     list_proxies()
 
@@ -192,17 +212,70 @@ def list_proxies():
     if not proxies:
         print("No proxies currently running.")
     else:
-        table = Table(title="VAC Proxies")
+        table = Table(title="VAC Proxies - `sunholo proxy list`")
         table.add_column("VAC")
         table.add_column("Port")
         table.add_column("PID")
         table.add_column("URL")
+        table.add_column("Local")
+        table.add_column("Logs")
         
         for service_name, info in proxies.items():
             url = f"http://127.0.0.1:{info['port']}"
-            table.add_row(service_name, str(info['port']), str(info['pid']), url)
+            table.add_row(service_name, 
+                          str(info['port']), 
+                          str(info['pid']), 
+                          url, 
+                          str(info['local']), 
+                          str(info['logs']) )
         
         console.print(table)
+
+def start_local(service_name, port, app_type, app_folder, log_file):
+    """
+    Starts a local Flask or FastAPI VAC app.
+
+    Args:
+        service_name (str): Name of the service.
+        port (int): Port to run the local app on.
+        app_type (str): Type of the app ('flask' or 'fastapi').
+        app_folder (str): Folder containing the app.
+    """
+    proxies = clean_proxy_list()
+
+    if service_name in proxies:
+        console.print(f"Local VAC app [bold orange]'{service_name}'[/bold orange] is already running on port {proxies[service_name]['port']}.")
+        return
+
+    if app_type == 'flask':
+        command = [sys.executable, 'app.py']
+    elif app_type == 'fastapi':
+        command = ["uvicorn", "app:app", f"--port={port}"]
+    else:
+        print(f"[bold red]Unknown app type: {app_type}[/bold red]")
+        return
+
+    if log_file:
+        log_file_path = os.path.join(app_folder, f"{service_name}_log.txt")
+        with open(log_file_path, 'w') as logf:
+            process = subprocess.Popen(command, cwd=app_folder, stdout=logf, stderr=logf, preexec_fn=os.setpgrp)
+    else:
+        log_file_path = "No log file specified"
+        with open(os.devnull, 'a') as devnull:
+            process = subprocess.Popen(command, cwd=app_folder, stdout=devnull, stderr=devnull, preexec_fn=os.setpgrp)
+
+    proxies[service_name] = {
+        "pid": process.pid,
+        "port": port,
+        "local": f"{app_folder}/app.py - {app_type}",
+        "logs": log_file_path
+    }
+    save_proxies(proxies)
+    
+    console.print(f"Local app [bold orange]'{service_name}'[/bold orange] started on port {port}")
+    list_proxies()
+
+    return f"http://127.0.0.1:{port}"
 
 def setup_proxy_subparser(subparsers):
     """
@@ -218,8 +291,19 @@ def setup_proxy_subparser(subparsers):
     start_parser = proxy_subparsers.add_parser('start', help='Start the proxy to the VAC Cloud Run service')
     start_parser.add_argument('service_name', help='Name of the Cloud Run service.')
     start_parser.add_argument('--port', type=int, help='Port to run the proxy on. Auto-assigns if not provided.')
-    start_parser.set_defaults(func=lambda args: start_proxy(args.service_name, args.region, args.project, args.port))
-
+    start_parser.add_argument('--local', action='store_true', help='Run the service locally instead of proxying to Cloud Run.')
+    start_parser.add_argument('--app-type', choices=['flask', 'fastapi'], help='If local, type of the local app (flask or fastapi).')
+    start_parser.add_argument('--app-folder', help='If local, folder containing the local app.py')
+    start_parser.add_argument('--log-file', action='store_true', help='Whether to create a file containing proxy logs.')
+    start_parser.set_defaults(func=lambda args: start_proxy(args.service_name, 
+                                                            args.region, 
+                                                            args.project, 
+                                                            args.port, 
+                                                            args.local, 
+                                                            args.app_type, 
+                                                            args.app_folder,
+                                                            args.log_file))
+    
     stop_parser = proxy_subparsers.add_parser('stop', help='Stop the proxy to the Cloud Run service.')
     stop_parser.add_argument('service_name', help='Name of the Cloud Run service.')
     stop_parser.set_defaults(func=lambda args: stop_proxy(args.service_name))
